requirements.txt
blinker==1.9.0
click==8.1.8
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==2.3.7
WTForms==3.2.1
cryptography
Flask-Migrate
Flask-CORS
email_validator
gunicorn==21.2.0
